Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
PyPDF2==3.0.1
google-generativeai==0.3.2